askcompany/requirements.txt
django~=3.0.11
pillow
django-debug-toolbar
django-extensions
django-bootstrap4